+++ packages/docs/stories/components/radio-group.stories.ts
@@ -7,7 +7,7 @@ import '../../../components/src/components/radio-group/radio-group.js';
 import type { Meta, StoryObj } from '@storybook/web-components';
 import { html } from 'lit';
 import { userEvent } from '@storybook/testing-library';
-import { generateStoryDescription, storybookDefaults } from '../../src/helpers/component.js';
+import { generateScreenshotStory, generateStoryDescription, storybookDefaults } from '../../src/helpers/component.js';
 
 const { args, argTypes } = storybookDefaults('syn-radio-group');
 
@@ -31,10 +31,12 @@ type Story = StoryObj;
 export const Default = {
   parameters: {
     controls: {
-      disable: true,
+      disable: false,
     },
     docs: {
-      description: generateStoryDescription('radio-group', 'default'),
+      description: {
+        story: generateStoryDescription('radio-group', 'default'),
+      },
     },
   },
   render: () => html`
@@ -46,12 +48,12 @@ export const Default = {
 } as Story;
 
 export const Labels: Story = {
+  name: 'Labels',
   parameters: {
-    controls: {
-      disable: true,
-    },
     docs: {
-      description: generateStoryDescription('radio-group', 'labels'),
+      description: {
+        story: generateStoryDescription('radio-group', 'labels'),
+      },
     },
   },
   render: () => html`
@@ -63,12 +65,12 @@ export const Labels: Story = {
 };
 
 export const HelpText: Story = {
+  name: 'Help text',
   parameters: {
-    controls: {
-      disable: true,
-    },
     docs: {
-      description: generateStoryDescription('radio-group', 'help-text'),
+      description: {
+        story: generateStoryDescription('radio-group', 'help-text'),
+      },
     },
   },
   render: () => html`
@@ -80,12 +82,12 @@ export const HelpText: Story = {
 };
 
 export const Disabled: Story = {
+  name: 'Disabled',
   parameters: {
-    controls: {
-      disable: true,
-    },
     docs: {
-      description: generateStoryDescription('radio-group', 'disabled'),
+      description: {
+        story: generateStoryDescription('radio-group', 'disabled'),
+      },
     },
   },
   render: () => html`
@@ -97,6 +99,7 @@ export const Disabled: Story = {
 };
 
 export const Checked: Story = {
+  name: 'Checked',
   render: () => html`
   <syn-radio-group label="This is a label" help-text="This is checked" name="a" value="2">
     <syn-radio value="1">Option</syn-radio>
@@ -106,12 +109,15 @@ export const Checked: Story = {
 };
 
 export const Invalid: Story = {
+  name: 'Invalid',
   parameters: {
-    controls: {
-      disable: true,
+    chromatic: {
+      disableSnapshot: false,
     },
     docs: {
-      description: generateStoryDescription('radio-group', 'required'),
+      description: {
+        story: generateStoryDescription('radio-group', 'invalid'),
+      },
     },
   },
   play: async ({ canvasElement }) => {
@@ -152,12 +158,15 @@ export const Invalid: Story = {
 };
 
 export const CustomValidity: Story = {
+  name: 'Custom validity',
   parameters: {
-    controls: {
-      disable: true,
+    chromatic: {
+      disableSnapshot: false,
     },
     docs: {
-      description: generateStoryDescription('radio-group', 'setCustomValidity'),
+      description: {
+        story: generateStoryDescription('radio-group', 'setCustomValidity'),
+      },
     },
   },
   play: async ({ canvasElement }) => {
@@ -199,7 +208,7 @@ export const CustomValidity: Story = {
   },
 
   render: () => html`
-  <form class="custom-validity">
+  <form>
     <syn-radio-group label="Select an option" name="a" value="1">
       <syn-radio value="1">Not me</syn-radio>
       <syn-radio value="2">Me neither</syn-radio>
@@ -209,3 +218,11 @@ export const CustomValidity: Story = {
     <syn-button type="submit" variant="filled">Submit</syn-button>
   </form>`,
 };
+
+// Bundled screenshot story
+export const Screenshot: Story = generateScreenshotStory([
+  Labels,
+  HelpText,
+  Disabled,
+  Checked,
+], 230);