+++ packages/tokens/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@synergy-design-system/tokens-v1.1.0](https://github.com/synergy-design-system/synergy-design-system/compare/tokens/1.0.3...tokens/1.1.0) (2024-01-11)
+
+
+### Features
+
+* ✨ add syn-icon-button ([#211](https://github.com/synergy-design-system/synergy-design-system/issues/211)) ([2a460be](https://github.com/synergy-design-system/synergy-design-system/commit/2a460be6bdad09c3a7b0fb211e92b26d71d5408e))
+
+# [@synergy-design-system/tokens-v1.0.3](https://github.com/synergy-design-system/synergy-design-system/compare/tokens/1.0.2...tokens/1.0.3) (2024-01-10)
+
+
+### Bug Fixes
+
+* 🤔 Replace primitive tokens of 'input.icon.icon-clearable' with core tokens ([#240](https://github.com/synergy-design-system/synergy-design-system/issues/240)) ([c2b8c18](https://github.com/synergy-design-system/synergy-design-system/commit/c2b8c18f6852c1abf95d2319730a321667435c0b))
+
+# [@synergy-design-system/tokens-v1.0.2](https://github.com/synergy-design-system/synergy-design-system/compare/tokens/1.0.1...tokens/1.0.2) (2023-12-14)
+
+
+### Bug Fixes
+
+* 📚 improve links in documentation ([#224](https://github.com/synergy-design-system/synergy-design-system/issues/224)) ([f55934c](https://github.com/synergy-design-system/synergy-design-system/commit/f55934c34c6c53b0f7c9a5afa8d91bc520df4fdf))
+
 # [@synergy-design-system/tokens-v1.0.1](https://github.com/synergy-design-system/synergy-design-system/compare/tokens/1.0.0...tokens/1.0.1) (2023-12-06)